refactor: remove ingest v1 version
