+++ indykite_sdk_node/src/version.ts
@@ -1,1 +1,1 @@
-export const LIB_VERSION = '0.4.2';
+export const LIB_VERSION = '0.4.4';
